requirements.txt
flask>=3.0.0
requests>=2.31.0
newsapi-python>=0.2.7
python-dotenv>=1.0.0
tabulate>=0.9.0
